dedup BuildLabels
