+++ internal/cri/server/podsandbox/helpers_test.go
@@ -19,37 +19,12 @@ package podsandbox
 import (
 	"context"
 	"os"
-	"strings"
 	"testing"
 
-	crilabels "github.com/containerd/containerd/v2/internal/cri/labels"
 	"github.com/containerd/containerd/v2/pkg/oci"
 	runtimespec "github.com/opencontainers/runtime-spec/specs-go"
 	"github.com/stretchr/testify/assert"
 )
-
-func TestBuildLabels(t *testing.T) {
-	imageConfigLabels := map[string]string{
-		"a":          "z",
-		"d":          "y",
-		"long-label": strings.Repeat("example", 10000),
-	}
-	configLabels := map[string]string{
-		"a": "b",
-		"c": "d",
-	}
-	newLabels := buildLabels(configLabels, imageConfigLabels, crilabels.ContainerKindSandbox)
-	assert.Len(t, newLabels, 4)
-	assert.Equal(t, "b", newLabels["a"])
-	assert.Equal(t, "d", newLabels["c"])
-	assert.Equal(t, "y", newLabels["d"])
-	assert.Equal(t, crilabels.ContainerKindSandbox, newLabels[crilabels.ContainerKindLabel])
-	assert.NotContains(t, newLabels, "long-label")
-
-	newLabels["a"] = "e"
-	assert.Empty(t, configLabels[crilabels.ContainerKindLabel], "should not add new labels into original label")
-	assert.Equal(t, "b", configLabels["a"], "change in new labels should not affect original label")
-}
 
 func TestEnvDeduplication(t *testing.T) {
 	for _, test := range []struct {
